Starting to create PositionCalculator module/service
What?
We tried to return the surrounding cell coords on a hash, but rubocop
was complaining about the ABC complexity of the method.
So what we're trying to do here is create a class by position that can
calculate that.

How?
We have a base class that makes everything basically, we just need to
inherit it to the position classes and define a method that calculates
X and Y.