Adding maximum x and y to PositionCalculator::Base constructor
What?
Until this point we didn't consider that we have a maximum X and Y.
We need them to know if the cell we're touching is on a corner or
an edge of the grid.

How?
We just need to add the params of max_x and max_y to the
PositionCalculator::Base constructor, and add those limits to the
nillify method.