Creating PositionCalculator::Factory
What?
Since we're creating a class for each position, we need a simple way
to instantiate the objects.

How?
For this we are creating a factory, that serves the position object
depending on the position name.

Also when creating the specs for this, we needed to require multiple
classes. This is kind of a pain in the ***. So we're including
zeitwerk for this.

Note: Next task has to be including zeitwerk on other specs