Install & set SimpleCov
What?
SimpleCov is a code coverage analysis tool for Ruby.
This will help us to know if we're making enough specs.
However this doesn't mean the specs are well done.

This is the github repo https://github.com/simplecov-ruby/simplecov.

How?
Adding the gem to the gemfile, setting up the initializer on the
rspec_helper file and ignoring the resulting HTML report.