Refactor Grid class, and adding GridObject Factory and Iterator
What?
Now that we have to make a Judge class, we need to make a lot of
grid styles to test. For this we need to make a GridObject Factory,
so it provides us with grids like "all_alive", "all_dead", "to_live",
"to_die" and "random".

How?
Created first a Support::Iterator. This consists of two methods
for the grid manipulation and iterations: .map and .each

After this, we create the Factory::GridObject. This will return a
matrix, that is the main grid thing.