Changing rspec_helper to spec_helper
What?
When we tried to run rspec --tag <tag>, we had errors. One of them
being that it works with a spec_helper.rb. We had it with a different
name.

We're also removing the Populator specs that was causing the command
to fail.

How?
Change the file name and all places where it is called.